+++ test_suite/foo.cpp
@@ -83,7 +83,7 @@ namespace test
 
 		void  PythonFoo ::testFooAutomaticFunctionExport( int iA, float iB )
 		{
-			LASS_COUT << "C++ testAutomaticFunctionExport : " << iA + iB << "\n";
+			LASS_COUT << "C++ testAutomaticFunctionExport : " << static_cast<float>(iA) + iB << "\n";
 		}
 	}
 }
